requirements.txt
RPi.GPIO
psutil